Create and fixing the uploaded file in create and edit project
